refactor: 3-tier hotkey architecture extraction
Extract the 800-line hotkeys.cc monolith (33 includes) into a layered
architecture to eliminate SPUD heisenbug risk:

Layer 1 - Thin hooks (hotkeys.cc, ~120 lines, 5 includes):
  SPUD detour hooks only, delegating all logic to the router.

Layer 2 - Dispatcher (hotkey_router.cc/h):
  Orchestration logic from ScreenManager_Update_Hook, routes to
  concern-specific functions.

Layer 3 - Concern files:
  - fleet_actions.cc/h: ship selection, ExecuteSpaceAction, recall/repair
  - viewer_mgmt.cc/h: object viewer visibility, rewards dismiss, info pending
  - cargo_display.cc/h: auto-show cargo on target scan
  - navigation.cc/h: section navigation, officer canvas

Also fixes implicit header dependencies in prime headers:
  - PreScanTargetWidget.h: add missing ObjectViewerBaseWidget.h and
    ScanEngageButtonsWidget.h includes
  - hotkey_dispatch.cc: use navigation.h instead of forward declarations
