+++ mods/src/patches/cargo_display.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class RewardsButtonWidget;
+struct PreScanTargetWidget;
+
+bool CheckShowCargo(RewardsButtonWidget* widget);
+void HandleCargoBindContext(RewardsButtonWidget* _this);
+void HandleCargoShowFleet(PreScanTargetWidget* _this);
