+++ mods/src/patches/viewer_mgmt.h
@@ -1,0 +1,8 @@
+#pragma once
+
+bool CanHideViewers();
+bool DidHideViewers();
+bool TryDismissRewardsScreen();
+void HandleActionView();
+void TickInfoPending();
+void SetInfoPending(int frames);
